+++ lib/n_plus_one_control/minitest.rb
@@ -9,7 +9,8 @@ module NPlusOneControl
       populate: nil,
       matching: nil,
       scale_factors: nil,
-      warmup: nil
+      warmup: nil,
+      collectors: :db
     )
 
       raise ArgumentError, "Block is required" unless block_given?
@@ -22,11 +23,16 @@ module NPlusOneControl
         scale_factors: scale_factors || NPlusOneControl.default_scale_factors
       )
 
-      queries = @executor.call { yield }
+      queries = @executor.call(collectors: collectors) { yield }
 
-      counts = queries.map(&:last).map(&:size)
+      counts = queries.map { |q| q.last.transform_values(&:size) }
 
-      assert counts.max == counts.min, NPlusOneControl.failure_message(:constant_queries, queries)
+      results = Array(collectors).map do |c|
+        counts_by_collector = counts.map { |count| count[c] }
+        [c, counts_by_collector.max == counts_by_collector.min]
+      end.to_h
+
+      assert results.values.all?, NPlusOneControl.failure_message(:constant_queries, queries)
     end
 
     def assert_perform_linear_number_of_queries(
@@ -34,7 +40,8 @@ module NPlusOneControl
       populate: nil,
       matching: nil,
       scale_factors: nil,
-      warmup: nil
+      warmup: nil,
+      collectors: :db
     )
 
       raise ArgumentError, "Block is required" unless block_given?
@@ -47,7 +54,7 @@ module NPlusOneControl
         scale_factors: scale_factors || NPlusOneControl.default_scale_factors
       )
 
-      queries = @executor.call { yield }
+      queries = @executor.call(collectors: collectors) { yield }
 
       assert linear?(queries, slope: slope), NPlusOneControl.failure_message(:linear_queries, queries)
     end
@@ -71,8 +78,11 @@ module NPlusOneControl
         scales = pair.map(&:first)
         query_lists = pair.map(&:last)
 
-        actual_slope = (query_lists[1].size - query_lists[0].size) / (scales[1] - scales[0])
-        actual_slope <= slope
+        actual_slopes = query_lists[0].keys.map do |key|
+          (query_lists[1][key].size - query_lists[0][key].size) / (scales[1] - scales[0])
+        end
+
+        actual_slopes.all? { |actual_slope| actual_slope <= slope }
       end
     end
   end