+++ tests/test_helper.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "minitest/autorun"
-require "minitest/pride"
 
 begin
   require "pry-byebug"